MULE-8278: HTTP Listener does not handle "Expect: 100-Continue" (RFC 2616) header correctly
Added test cases for proxy scenario
