+++ modules/http/src/test/java/org/mule/module/http/functional/requester/HttpRequestExpectHeaderTestCase.java
@@ -14,36 +14,18 @@ import static org.mule.module.http.api.HttpHeaders.Names.EXPECT;
 import static org.mule.module.http.api.HttpHeaders.Values.CONTINUE;
 import org.mule.api.MuleEvent;
 import org.mule.construct.Flow;
-import org.mule.tck.junit4.FunctionalTestCase;
-import org.mule.tck.junit4.rule.DynamicPort;
-import org.mule.util.IOUtils;
-import org.mule.util.concurrent.Latch;
+import org.mule.module.http.functional.AbstractHttpExpectHeaderServerTestCase;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.net.ServerSocket;
-import java.net.Socket;
 
-import org.junit.Rule;
 import org.junit.Test;
 
 
-public class HttpRequestExpectHeaderTestCase extends FunctionalTestCase
+public class HttpRequestExpectHeaderTestCase extends AbstractHttpExpectHeaderServerTestCase
 {
 
-    private static final String CONTINUE_RESPONSE = "HTTP/1.1 100 Continue\r\n\r\n";
-    private static final String EXPECTATION_FAILED_RESPONSE = "HTTP/1.1 417 Expectation Failed\r\n";
     private static final String REQUEST_FLOW_NAME = "requestFlow";
-
-    @Rule
-    public DynamicPort listenPort = new DynamicPort("httpPort");
-
-    private String requestBody;
 
     @Override
     protected String getConfigFile()
@@ -54,8 +36,7 @@ public class HttpRequestExpectHeaderTestCase extends FunctionalTestCase
     @Test
     public void handlesContinueResponse() throws Exception
     {
-        ExpectContinueMockServer server = new ExpectContinueMockServer();
-        server.start();
+        startExpectContinueServer();
 
         Flow flow = (Flow) getFlowConstruct(REQUEST_FLOW_NAME);
         MuleEvent event = getTestEvent(TEST_MESSAGE);
@@ -64,14 +45,13 @@ public class HttpRequestExpectHeaderTestCase extends FunctionalTestCase
         flow.process(event);
         assertThat(requestBody, equalTo(TEST_MESSAGE));
 
-        server.stop();
+        stopServer();
     }
 
     @Test
     public void handlesExpectationFailedResponse() throws Exception
     {
-        ExpectFailedMockServer server = new ExpectFailedMockServer();
-        server.start();
+        startExpectFailedServer();
 
         Flow flow = (Flow) getFlowConstruct(REQUEST_FLOW_NAME);
 
@@ -92,110 +72,7 @@ public class HttpRequestExpectHeaderTestCase extends FunctionalTestCase
         assertThat(response.getMessage().<Integer>getInboundProperty(HTTP_STATUS_PROPERTY),
                    equalTo(EXPECTATION_FAILED.getStatusCode()));
 
-        server.stop();
+        stopServer();
     }
 
-    private abstract class AbstractMockServer implements Runnable
-    {
-
-        private Latch startedLatch = new Latch();
-        private Latch finishedLatch = new Latch();
-
-        public void start()
-        {
-            try
-            {
-                Thread serverThread = new Thread(this);
-                serverThread.start();
-                startedLatch.await();
-            }
-            catch (InterruptedException e)
-            {
-                throw new RuntimeException("Failed to start mock server", e);
-            }
-        }
-
-        public void stop()
-        {
-            try
-            {
-                finishedLatch.await();
-            }
-            catch (InterruptedException e)
-            {
-                throw new RuntimeException("Failed to stop mock server", e);
-            }
-        }
-
-        @Override
-        public void run()
-        {
-            try
-            {
-                ServerSocket serverSocket = new ServerSocket(listenPort.getNumber());
-                startedLatch.release();
-                Socket socket = serverSocket.accept();
-                InputStream inputStream = socket.getInputStream();
-                OutputStream outputStream = socket.getOutputStream();
-
-                BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream), 1);
-                BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(outputStream), 1);
-
-                while (!reader.readLine().isEmpty())
-                {
-                    // Do nothing, consume headers until blank line.
-                }
-
-                process(reader, writer);
-
-                reader.close();
-                writer.close();
-                socket.close();
-                serverSocket.close();
-
-                finishedLatch.release();
-            }
-            catch (IOException e)
-            {
-                throw new RuntimeException(e);
-            }
-        }
-
-        protected abstract void process(BufferedReader reader, BufferedWriter writer) throws IOException;
-    }
-
-
-    private class ExpectContinueMockServer extends AbstractMockServer
-    {
-
-        @Override
-        protected void process(BufferedReader reader, BufferedWriter writer) throws IOException
-        {
-            writer.write(CONTINUE_RESPONSE);
-            writer.flush();
-
-            char[] body = new char[TEST_MESSAGE.length()];
-            IOUtils.read(reader, body);
-            requestBody = new String(body);
-
-            String response = "HTTP/1.1 200 OK\n" +
-                              "Content-Length: 4\n\n" +
-                              "TEST";
-
-            writer.write(response);
-            writer.flush();
-        }
-    }
-
-
-    private class ExpectFailedMockServer extends AbstractMockServer
-    {
-
-        @Override
-        protected void process(BufferedReader reader, BufferedWriter writer) throws IOException
-        {
-            writer.write(EXPECTATION_FAILED_RESPONSE);
-            writer.flush();
-        }
-    }
 }
